requirements.txt
asgiref==3.7.2
cachetools==4.2.4
certifi==2023.11.17
charset-normalizer==3.3.2
Django==4.1.12
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
google-auth==2.0.1
idna==3.6
inflection==0.5.1
packaging==23.2
Pillow==10.1.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
sqlparse==0.4.4
uritemplate==4.1.1
urllib3==2.1.0
